Rebuild "motion" fonts per 71a6883, 94b6450, 7280e5d
As with commits d0f38b6, 7f80ee6, this does not yet form part of a
release.  I am testing all the fonts to ensure everything works as
intended.  I will prepare a release when the time is right.

For the underlying reason, refer to the commit message of 7f80ee6.